fix: silence MSVC C4146/C4113 warnings
- regex.c: cast len to int64_t before unary minus so the comparison
  is a proper signed compare instead of unsigned wrap (C4146).
  The branch is currently unreachable due to an earlier early-return;
  tracked separately in #499.
- dirmonitor backends: use (void) instead of K&R () in
  get_mode_dirmonitor declarations so MSVC treats them as
  matching the function pointer in dirmonitor_backend (C4113).

+++ src/tokenizer/regex.c
@@ -157,7 +157,7 @@ static size_t regex_offset_relative(int64_t pos, size_t len) {
     return pos;
   else if (pos == 0)
     return 1;
-  else if (pos < -len)  /* inverted comparison */
+  else if (pos < -(int64_t)len)  /* inverted comparison */
     return 1;  /* clip to 1 */
   else return len + pos + 1;
 }
